move modTypes -> types/
