+++ src/extensionConsole/pages/mods/emptyView/EmptyView.tsx
@@ -17,7 +17,7 @@
 
 import React from "react";
 import styles from "./EmptyView.module.scss";
-import { type ModViewItem } from "@/mods/modTypes";
+import { type ModViewItem } from "@/types/modTypes";
 import { type TableInstance } from "react-table";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faExternalLinkAlt } from "@fortawesome/free-solid-svg-icons";
