+++ src/extensionConsole/pages/mods/labels/SharingLabel.tsx
@@ -26,7 +26,7 @@ import {
 } from "@fortawesome/free-solid-svg-icons";
 import { OverlayTrigger, Popover } from "react-bootstrap";
 import cx from "classnames";
-import { type SharingSource } from "@/mods/modTypes";
+import { type SharingSource } from "@/types/modTypes";
 
 const sharingIcons = {
   Personal: faEyeSlash,
