+++ src/mods/ModIcon.tsx
@@ -24,7 +24,7 @@ import {
 } from "@fortawesome/free-solid-svg-icons";
 import { useAsyncIcon } from "@/components/asyncIcon";
 import { type MarketplaceListing } from "@/types/contract";
-import { type Mod } from "@/mods/modTypes";
+import { type Mod } from "@/types/modTypes";
 import {
   getPackageId,
   isModDefinition,
