+++ src/mods/hooks/useRequestPermissionsAction.ts
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { type ModViewItem } from "@/mods/modTypes";
+import { type ModViewItem } from "@/types/modTypes";
 import { useCallback } from "react";
 import { type OptionsState } from "@/store/extensionsTypes";
 import { selectExtensionsFromMod } from "@/utils/modUtils";
